requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary==2.9.9
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
python-dotenv
anthropic
pydantic[email]
PyPDF2
Pillow
emails
aiofiles
jinja2
httpx
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
alembic